requirements.txt
# Core LLM & Embeddings
google-generativeai>=0.7.0
google-cloud-aiplatform>=1.50.0

# Document Parsing
pymupdf>=1.24.0
unstructured[pdf,image]>=0.13.0
camelot-py[cv]>=0.11.0
pillow>=10.0.0
pytesseract>=0.3.10

# Vector Store
qdrant-client>=1.9.0

# RAG Framework
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0

# API
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.0
sse-starlette>=2.1.0

# UI
streamlit>=1.35.0
streamlit-chat>=0.1.1

# Database (feedback loop)
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
alembic>=1.13.0

# Fine-tuning / RLHF
transformers>=4.41.0
trl>=0.9.0
datasets>=2.19.0
peft>=0.11.0
accelerate>=0.30.0

# Evaluation
ragas>=0.1.9
pandas>=2.2.0
numpy>=1.26.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tenacity>=8.3.0
httpx>=0.27.0
tqdm>=4.66.0
